+++ src/api/endpoints.tsx
@@ -1,7 +1,6 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
 import {
-    DetailedDish,
-    DetailedDishWithTranslations,
+    DetailedDishWithTranslations, DishDocument,
     DishProposal,
     DishResult,
     NotActivatedUser
@@ -9,10 +8,7 @@ import {
 import {
     getDish, getDishComments,
     getDishProposals,
-    getDishes, getNotActivatedUsers,
-    getSoftAddedDishes,
-    getSoftDeletedDishes,
-    getSoftEditedDishes
+    getDishes, getNotActivatedUsers, getSoftAddedDishes, getSoftEditedDishes, getSoftDeletedDishes,
 } from '@/src/api/api';
 import { DAY, HOUR, MINUTE } from '@/src/constants/numbers';
 import { DishComment } from '@/src/types/dish.types';
@@ -42,7 +38,7 @@ export function useGetDishProposals(): UseQueryResult<DishProposal[]> {
 }
 
 // TODO: Use to fetch data. Trigger hooks using boolean value
-export function useGetSoftAddedDishes(isTriggered: boolean = true): UseQueryResult<DetailedDish[]> {
+export function useGetSoftAddedDishes(isTriggered: boolean = true): UseQueryResult<DishDocument[]> {
     if (!isTriggered) {
         return useQuery({
             queryKey: ['dishes', 'soft', 'added'],
@@ -51,14 +47,14 @@ export function useGetSoftAddedDishes(isTriggered: boolean = true): UseQueryResu
     }
 
     return useQuery({
-        queryFn: async (): Promise<DetailedDish[]> => await getSoftAddedDishes(),
+        queryFn: async (): Promise<DishDocument[]> => await getSoftAddedDishes(),
         queryKey: ['dishes', 'soft', 'added'],
         staleTime: HOUR
     });
 }
 
 // TODO: Use to fetch data. Trigger hooks using boolean value
-export function useGetSoftEditedDishes(isTriggered: boolean = true): UseQueryResult<DetailedDish[]> {
+export function useGetSoftEditedDishes(isTriggered: boolean = true): UseQueryResult<DishDocument[]> {
     if (!isTriggered) {
         return useQuery({
             queryKey: ['dishes', 'soft', 'edited'],
@@ -67,14 +63,14 @@ export function useGetSoftEditedDishes(isTriggered: boolean = true): UseQueryRes
     }
 
     return useQuery({
-        queryFn: async (): Promise<DetailedDish[]> => await getSoftEditedDishes(),
+        queryFn: async (): Promise<DishDocument[]> => await getSoftEditedDishes(),
         queryKey: ['dishes', 'soft', 'edited'],
         staleTime: HOUR
     });
 }
 
 // TODO: Use to fetch data. Trigger hooks using boolean value
-export function useGetSoftDeletedDishes(isTriggered: boolean = true): UseQueryResult<DetailedDish[]> {
+export function useGetSoftDeletedDishes(isTriggered: boolean = true): UseQueryResult<DishDocument[]> {
     if (!isTriggered) {
         return useQuery({
             queryKey: ['dishes', 'soft', 'deleted'],
@@ -83,7 +79,7 @@ export function useGetSoftDeletedDishes(isTriggered: boolean = true): UseQueryRe
     }
 
     return useQuery({
-        queryFn: async (): Promise<DetailedDish[]> => await getSoftDeletedDishes(),
+        queryFn: async (): Promise<DishDocument[]> => await getSoftDeletedDishes(),
         queryKey: ['dishes', 'soft', 'deleted'],
         staleTime: HOUR
     });