+++ src/app/users/registration/page.tsx
@@ -4,13 +4,9 @@ import { WrappedScreenContentLayout } from '@/src/components/common/layouts/wrap
 
 export default function RegistrationPage() {
     return (
-        <WrappedScreenContentLayout>
-            <div>
-                <div style={{ margin: '.5rem' }}>
-                    <BackLinkBar link="/users/login" label={'Powrót do logowania'} />
-                </div>
-                <RegistrationForm />
-            </div>
+        <WrappedScreenContentLayout type="multi-content">
+            <BackLinkBar link="/users/login" label={'Powrót do logowania'} />
+            <RegistrationForm />
         </WrappedScreenContentLayout>
     );
 }